requirements.txt
streamlit
qdrant-client
google-generativeai
python-dotenv
langchain-text-splitters
pypdf
python-docx